+++ validated/src/main/java/uk/ac/standrews/cs/digitising_scotland/verisim/populationStatistics/dataDistributionTables/statsKeys/MultipleBirthStatsKey.java
@@ -32,7 +32,7 @@ public class MultipleBirthStatsKey extends StatsKey {
         return getYLabel();
     }
 
-    public int getForNumberOfChildren() {
+    public Double getForNumberOfChildren() {
         return getForNPeople();
     }
