+++ validated/src/main/java/uk/ac/standrews/cs/digitising_scotland/verisim/populationStatistics/dataDistributionTables/statsKeys/StatsKey.java
@@ -27,11 +27,11 @@ public class StatsKey {
     private final Integer yLabel;
     private final Integer xLabel;
     private final Integer maxXLabel;
-    private final int forNPeople;
+    private final double forNPeople;
     private final CompoundTimeUnit consideredTimePeriod;
     private final Date date;
 
-    public StatsKey(Integer yLabel, Integer xLabel, Integer maxXLabel, int forNPeople, CompoundTimeUnit consideredTimePeriod, Date date) {
+    public StatsKey(Integer yLabel, Integer xLabel, Integer maxXLabel, double forNPeople, CompoundTimeUnit consideredTimePeriod, Date date) {
         this.yLabel = yLabel;
         this.xLabel = xLabel;
         this.maxXLabel = maxXLabel;
@@ -40,15 +40,15 @@ public class StatsKey {
         this.date = date;
     }
 
-    public StatsKey(Integer yLabel, Integer xLabel, int forNPeople, CompoundTimeUnit consideredTimePeriod, Date date) {
+    public StatsKey(Integer yLabel, Integer xLabel, double forNPeople, CompoundTimeUnit consideredTimePeriod, Date date) {
         this(yLabel, xLabel, null, forNPeople, consideredTimePeriod, date);
     }
 
-    public StatsKey(Integer yLabel, int forNPeople, CompoundTimeUnit consideredTimePeriod, Date date) {
+    public StatsKey(Integer yLabel, double forNPeople, CompoundTimeUnit consideredTimePeriod, Date date) {
         this(yLabel, null, null, forNPeople, consideredTimePeriod, date);
     }
 
-    public int getForNPeople() {
+    public double getForNPeople() {
         return forNPeople;
     }
